requirements.txt
joblib
requests
scikit-learn==1.7.1
pandas
boto3
duckdb==1.2.2
python-dotenv
seaborn
matplotlib
numpy
Flask
waitress
flask-cors
Flask-Swagger-UI
asset-model-data-storage
tensorflow
